+++ templates/answer.html
@@ -11,6 +11,13 @@
 <div style="display: flex; align-items: center; justify-content: center">
 <h1>!תודה שהשתמשת באתר שלנו</h1>
 </div>
+<div style="display: flex; align-items: center; justify-content: center">
+<a href="/previous_page?algo={{num_algo}}" style="font-size: 20px">חזור</a>
+</div>
+         <br>
+<div style="display: flex; align-items: center; justify-content: center">
+<a href="/back_home_page" style="font-size: 20px">לדף הבית</a>
+</div>
 </body>
 </html>
